separated stuff into components, wip
