+++ app/javascript/googlemap.vue
@@ -1,0 +1,25 @@
+<template>
+  <div>
+    <gmap-map
+      :center="center"
+      :zoom="12"
+      style="width:100%;  height: 400px;"
+    >
+      <gmap-marker
+        :key="index"
+        v-for="(m, index) in markers"
+        :position="m.position"
+        @click="center=m.position"
+      ></gmap-marker>
+    </gmap-map>
+  </div>
+</template>
+
+<script>
+  export default {
+    name: "GoogleMap",
+    data: {
+      potholes: []
+    }
+  }
+</script>
