+++ app/views/datas/index.html.erb
@@ -1,4 +1,3 @@
 <div id='app'>
-     {{ message }}
   <app></app>
 </div>
