+++ package.json
@@ -7,7 +7,8 @@
     "vue": "^2.5.16",
     "vue-axios": "^2.1.1",
     "vue-loader": "14.2.2",
-    "vue-template-compiler": "^2.5.16"
+    "vue-template-compiler": "^2.5.16",
+    "vue2-google-maps": "^0.9.7"
   },
   "devDependencies": {
     "webpack-dev-server": "2.11.2"
